# covid-cluster-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "covid-cluster",
    "contact": {}
  },
  "host": "localhost:9000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/covidCluster": {
      "post": {
        "summary": "Add covid cluster",
        "tags": [
          "Misc"
        ],
        "operationId": "Addcovidcluster",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AddcovidclusterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete Cluster",
        "tags": [
          "Misc"
        ],
        "operationId": "DeleteCluster",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/DeleteClusterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
...
  },
  "definitions": {
    "AddcovidclusterRequest": {
      "title": "AddcovidclusterRequest",
      "example": {
        "baseClusterName": "zero",
        "baseClusterSide": 0,
        "newClusterName": "zero-zero"
      },
      "type": "object",
      "properties": {
        "baseClusterName": {
          "type": "string"
        },
        "baseClusterSide": {
          "type": "integer",
          "format": "int32"
        },
        "newClusterName": {
          "type": "string"
        }
      },
      "required": [
        "baseClusterName",
        "baseClusterSide",
        "newClusterName"
      ]
    },
    "DeleteClusterRequest": {
      "title": "DeleteClusterRequest",
      "example": {
        "clusterName": "origin"
      },
      "type": "object",
      "properties": {
        "clusterName": {
          "type": "string"
        }
      },
      "required": [
        "clusterName"
      ]
    }
  },
...
}